backend/requirements.txt
# Backend API (FastAPI)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
python-dotenv>=1.0.0
pyjwt>=2.9.0
